--- apps/fastworld-marketplace-backend/src/categories/categories.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import CategoriesService  from './categories.service';
-import CategoriesController from './categories.controller';
-import Category from './entitys/category.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Category])],
-  providers: [CategoriesService],
-  controllers: [CategoriesController],
-})
-export class CategoriesModule {}
