--- apps/fastworld-marketplace-backend/src/products/products.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import Product from '../../../../product/src/entitys/products.entity';
-import ProductsController from './products.controller';
-import ProductService from './products.service';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Product])],
-  controllers: [ProductsController],
-  providers: [ProductService],
-})
-export default class ProductsModule {}
