+++ apps/fastworld-marketplace/src/app.controller.ts
@@ -1,9 +1,9 @@
 import { Controller, Get } from '@nestjs/common';
-import AppService from './app.service';
+import { AppService } from './app.service';
 
 @Controller()
-export default class AppController {
+export class AppController {
   constructor(
     private readonly appService: AppService
-  ) {}
+  ) { }
 }
