+++ apps/fastworld-marketplace/src/app.service.ts
@@ -3,6 +3,5 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class AppService {
   constructor(
-
   ){}
 }
