+++ apps/fastworld-marketplace/src/modules/users/modules/profiles/profiles.module.ts
@@ -1,13 +1,11 @@
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProfilesController } from './profiles.controller';
 import { ProfilesService } from './profiles.service';
 import { Module } from '@nestjs/common';
-import Profile from './entitys/profile.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Profile])],
+  imports: [],
   controllers: [ProfilesController],
   providers: [ProfilesService],
   exports: [ProfilesService],
 })
-export default class ProfilesModule {}
+export class ProfilesModule {}
