+++ apps/fastworld-marketplace/src/modules/users/modules/role/roles.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { RoleService } from './roles.service';
+import { RoleController } from './roles.controller';
+
+@Module({
+  imports: [],
+  controllers: [RoleController],
+  providers: [RoleService],
+  exports: [RoleService],
+})
+export class RoleModule {}
